renamed top level folder from AppBar to NavBar
