+++ src/components/NavBar/NavBarWithSearch/NavBarWithSearch.tsx
